+++ CHANGELOG.md
@@ -3,6 +3,9 @@ yyyy/mm/dd Version x.y.z
 
 - Add response `timeout` option with 250ms default and per-method overrides
 - Add response `retry_times` option with 0 default and per-method overrides
+- Default port is now `/dev/ttyUSB0`
+- Renamed `crystalfontz atx` arg `--power-pulse-length-seconds` to `--power-pulse-length`
+- Improved CLI help text
 
 2025/01/06 Version 2.0.0
 ------------------------